Max 20 canvases on screen: extract Canvas generator
In preparation to limit Canvases, the commit extracts a function
responsible for thier preparation given vector of peer indetifiers.
No changes in canvas preparation logic were introduced.
