+++ yew-ui/src/components/mod.rs
@@ -3,4 +3,6 @@ pub mod device_selector;
 pub mod host;
 pub mod icons;
 pub mod top_bar;
+
+mod canvas_generator;
 mod peer_list;
